requirements.txt
openai
requests
moviepy==1.0.3
pillow==9.5.0
python-dotenv
